+++ spec/components/summary_list_spec.rb
@@ -15,7 +15,7 @@ describe "Summary list", type: :view do
     assert_select ".gem-c-summary-list .govuk-heading-m", text: "Title, summary and body"
   end
 
-  it "renders section title with edit action" do
+  it "renders section title with edit and delete actions" do
     render_component(
       title: "Title, summary and body",
       edit: {
@@ -24,21 +24,30 @@ describe "Summary list", type: :view do
           gtm: "edit-title-summary-body",
         },
       },
+      delete: {
+        href: "#delete-title-summary-body",
+        data_attributes: {
+          gtm: "delete-title-summary-body",
+        },
+      },
     )
-    assert_select ".gem-c-summary-list .govuk-heading-m", text: "Title, summary and body"
-    assert_select '.gem-c-summary-list__edit-section-link[title="Change Title, summary and body"][href="#edit-title-summary-body"][data-gtm="edit-title-summary-body"]', text: "Change Title, summary and body"
+    assert_select ".gem-c-summary-list h3.govuk-heading-m.gem-c-summary-list__group-title", text: "Title, summary and body"
+    assert_select '.gem-c-summary-list__group-actions-list .govuk-link[title="Change Title, summary and body"][href="#edit-title-summary-body"][data-gtm="edit-title-summary-body"]', text: "Change Title, summary and body"
+    assert_select '.gem-c-summary-list__group-actions-list .govuk-link.gem-link--destructive[title="Delete Title, summary and body"][href="#delete-title-summary-body"][data-gtm="delete-title-summary-body"]', text: "Delete Title, summary and body"
   end
 
-  it "renders section title with custom link text" do
+  it "renders section title with custom link text and heading level" do
     render_component(
       title: "Items",
+      heading_level: 2,
+      heading_size: "l",
       edit: {
         href: "#custom-action",
         link_text: "Reorder",
       },
     )
-    assert_select ".gem-c-summary-list .govuk-heading-m", text: "Items"
-    assert_select '.gem-c-summary-list__edit-section-link[title="Reorder Items"][href="#custom-action"]', text: "Reorder Items"
+    assert_select ".gem-c-summary-list h2.govuk-heading-l", text: "Items"
+    assert_select '.gem-c-summary-list__group-actions-list .govuk-link[title="Reorder Items"][href="#custom-action"]', text: "Reorder Items"
   end
 
   it "renders section title with block" do
@@ -113,7 +122,7 @@ describe "Summary list", type: :view do
     )
     assert_select ".govuk-summary-list__key", text: "Title"
     assert_select ".govuk-summary-list__value", text: "Ethical standards for public service providers"
-    assert_select '.govuk-summary-list__actions-list-item .govuk-link[title="Delete Title"][href="#delete-title"][data-gtm="delete-title"]', text: "Delete Title"
+    assert_select '.govuk-summary-list__actions-list-item .govuk-link.gem-link--destructive[title="Delete Title"][href="#delete-title"][data-gtm="delete-title"]', text: "Delete Title"
   end
 
   it "renders items with custom text for edit and delete action" do
@@ -134,6 +143,6 @@ describe "Summary list", type: :view do
       ],
     )
     assert_select '.govuk-summary-list__actions-list-item .govuk-link[title="Edit Title"][href="#edit-title"]', text: "Edit Title"
-    assert_select '.govuk-summary-list__actions-list-item .govuk-link[title="Remove Title"][href="#delete-title"]', text: "Remove Title"
+    assert_select '.govuk-summary-list__actions-list-item .govuk-link.gem-link--destructive[title="Remove Title"][href="#delete-title"]', text: "Remove Title"
   end
 end